### 2025-08-14 #8790 @jmorris335
**Editor suggestions:**

The paper Psymple, a software with similar use cases as ConstraintHg recently published in JOSS. Any of developers of this software would be excellent choices for reviewers. They also have some familiarity with ConstraintHg.
- George Simmons ([GJHSimmons](https://github.com/GJHSimmons)), main developer
- George Osange ([geoo89](https://github.com/geoo89))
- Tim Lusgood ([thosgood](https://github.com/geoo89)), more focused on the mathematics

Other JOSS reviewers that appear to have relevant backgrounds in Python and systems modeling:
- [Fergus Baker](https://reviewers.joss.theoj.org/reviewers/21559) ([fjebaker](https://github.com/fjebaker))
- [Milan Rother](https://reviewers.joss.theoj.org/reviewers/15882) ([milanofthe](https://github.com/milanofthe))
- [Ewout ter Hoeven](https://reviewers.joss.theoj.org/reviewers/8753) ([EwoutH](https://github.com/EwoutH))
- [Samuel Macharia Kang'iri](https://reviewers.joss.theoj.org/reviewers/3832) ([SamMachariaPhD](https://github.com/SamMachariaPhD))
- [William Schueller](https://reviewers.joss.theoj.org/reviewers/19950) ([wschuell](https://github.com/wschuell))
