### 2025-08-14 bd8f932
Updated missing DOIs
