+++ paper/paper.bib
@@ -23,10 +23,11 @@
   publisher = {Open Publishing Association},
   location = {Glasgow, United Kingdom},
   doi = {10.4204/EPTCS.380.5},
+  abstract = {Stock and flow diagrams are widely used in epidemiology to model the dynamics of populations. Although tools already exist for building these diagrams and simulating the systems they describe, we have created a new package called StockFlow, part of the AlgebraicJulia ecosystem, which uses ideas from category theory to overcome notable limitations of existing software. Compositionality is provided by the theory of decorated cospans: stock and flow diagrams can be composed to form larger ones in an intuitive way formalized by the operad of undirected wiring diagrams. Our approach also cleanly separates the syntax of stock and flow diagrams from the semantics they can be assigned. We consider semantics in ordinary differential equations, although others are possible. As an example, we explain code in StockFlow that implements a simplified version of a COVID-19 model used in Canada.},
   eventtitle = {Fifth {{International Conference}} on {{Applied Category Theory}} ({{ACT2023}})},
+  keywords = {Computer Science - Logic in Computer Science,Computer Science - Mathematical Software,Mathematics - Category Theory,Quantitative Biology - Populations and Evolution},
   annotation = {EPTCS: 380.5}
 }
-
 
 @incollection{baezPhysicsTopologyLogic2010,
   title = {Physics, {{Topology}}, {{Logic}} and {{Computation}}: {{A Rosetta Stone}}},
@@ -40,8 +41,10 @@
   publisher = {Springer Berlin Heidelberg},
   location = {Berlin, Heidelberg},
   doi = {10.1007/978-3-642-12821-9_2},
+  abstract = {In physics, Feynman diagrams are used to reason about quantum processes. In the 1980s, it became clear that underlying these diagrams is a powerful analogy between quantum physics and topology. Namely, a linear operator behaves very much like a ‘cobordism’: a manifold representing spacetime, going between two manifolds representing space. This led to a burst of work on topological quantum field theory and ‘quantum topology’. But this was just the beginning: similar diagrams can be used to reason about logic, where they represent proofs, and computation, where they represent programs. With the rise of interest in quantum cryptography and quantum computation, it became clear that there is extensive network of analogies between physics, topology, logic and computation. In this expository paper, we make some of these analogies precise using the concept of ‘closed symmetric monoidal category’. We assume no prior knowledge of category theory, proof theory or computer science.},
   isbn = {978-3-642-12820-2 978-3-642-12821-9},
   langid = {english},
+  keywords = {TOREAD,USEFUL}
 }
 
 @book{borutzkyBondGraphModelling2011,
@@ -56,20 +59,6 @@
   langid = {english}
 }
 
-@article{breinerCategoriesPlanningScheduling2020,
-  title = {Categories for Planning and Scheduling},
-  author = {Breiner, Spencer and Denno, Peter O. and Subrahmanian, Eswaran},
-  date = {2020-12-01T07:00-05:00},
-  journaltitle = {NIST},
-  volume = {67},
-  number = {11},
-  publisher = {Spencer Breiner, Peter O. Denno, Eswaran Subrahmanian},
-  url = {https://www.nist.gov/publications/categories-planning-and-scheduling},
-  urldate = {2025-02-25},
-  abstract = {N/A},
-  langid = {english},
-}
-
 @book{cellierContinuousSystemModeling1991,
   title = {Continuous {{System Modeling}}},
   author = {Cellier, François E.},
@@ -77,9 +66,11 @@
   publisher = {Springer},
   location = {New York, NY},
   doi = {10.1007/978-1-4757-3922-0},
[…]
