### 2025-09-09 c988c6e
Added hg.print_nodes()
