### 2025-09-19 f585439
Overhauled documentation
