+++ docs/source/constrainthg.rst
@@ -1,57 +1,0 @@
-ConstraintHg Package
-====================
-
-Hypergraph Module
-------------------------------
-
-Tree Node Class
-***********************************
-.. autoclass:: constrainthg.hypergraph.TNode
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Node Class
-***********************************
-.. autoclass:: constrainthg.hypergraph.Node
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-EdgeProperty Class
-***********************************
-.. autoclass:: constrainthg.hypergraph.EdgeProperty
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Edge Class
-***********************************
-.. autoclass:: constrainthg.hypergraph.Edge
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Pathfinder Class
-***********************************
-.. autoclass:: constrainthg.hypergraph.Pathfinder
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Hypergraph Class
-***********************************
-.. autoclass:: constrainthg.hypergraph.Hypergraph
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Relations Module
------------------------------
-
-.. automodule:: constrainthg.relations
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-:doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
