+++ docs/source/api/edge.rst
@@ -1,0 +1,18 @@
+Edge
+====
+
+EdgeProperty Class
+***********************************
+.. autoclass:: constrainthg.hypergraph.EdgeProperty
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Edge Class
+***********************************
+.. autoclass:: constrainthg.hypergraph.Edge
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+:doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
