+++ docs/source/chg_intro.rst
@@ -51,7 +51,7 @@ However, we can describe more relationships in the system. For instance, if we k
 
 Take note that the :math:`\alpha` node in :ref:`Figure 3 <pend_damping>` has two edges leading to it, showing that there are two possible ways to constrain its values. This is a quirk of constraint hypergraphs that permits competing constraints–a normal constraint network would not allow a variable to be related by two different, overlapping constraints. But in this case by allowing the dual relationships we have indicated that there are at least two possible models that can be used to solve for :math:`\alpha` (in this case one with and one without damping). During simulation, if we have a damping coefficient, the solver has to select which model will be active at any given time. 
 
-The takeaway is that the constraint hypergraph shows not just a single way to simulate the system, it shows *every* way to simulate the system—every possible model we can construct [#]_. We could add more edges and nodes to the graph if we knew relationships between :math:`\alpha` and :math:`\omega`, or :math:`\omega` and :math:`\theta`, or any other variable we want to virtually observe. The constraint hypergraph captures them all and provides the solver the ability to distinguish between them, chain edges together (path finding), and select preferred models. For a full implementation see the `Pendulum demo <https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_pendulum.py>` in the package.
+The takeaway is that the constraint hypergraph shows not just a single way to simulate the system, it shows *every* way to simulate the system—every possible model we can construct [#]_. We could add more edges and nodes to the graph if we knew relationships between :math:`\alpha` and :math:`\omega`, or :math:`\omega` and :math:`\theta`, or any other variable we want to virtually observe. The constraint hypergraph captures them all and provides the solver the ability to distinguish between them, chain edges together (path finding), and select preferred models. For a full implementation see the `Pendulum demo <https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_pendulum.py>`_ in the package.
 
 What are they used for
 ======================
