+++ docs/source/api/hypergraph.rst
@@ -1,0 +1,18 @@
+Hypergraph Module
+=================
+
+Pathfinder Class
+***********************************
+.. autoclass:: constrainthg.hypergraph.Pathfinder
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Hypergraph Class
+***********************************
+.. autoclass:: constrainthg.hypergraph.Hypergraph
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+:doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
