+++ docs/source/demos.rst
@@ -1,0 +1,19 @@
+Demonstrations
+==============
+
+.. toctree::
+    :hidden:
+    
+    Simple Demo <simple_demo>
+    Pendulum <https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_pendulum.py>
+    Microgrid <https://github.com/jmorris335/MicrogridHg>
+    Crankshaft <https://github.com/jmorris335/tool-interoperability-scripts/tree/main>
+    3D Printer <https://github.com/jmorris335/Powder-Based-Fusion-Digital-Twin>
+    Elevator <https://github.com/jmorris335/ElevatorHypergraph>
+
+.. include:: chg_overview.rst
+    :start-after: .. start demonstrations
+    :end-before: .. end demonstrations
+
+:ref:`genindex` \| :ref:`modindex` \| :ref:`Search <search>`
+
