+++ docs/source/api/relations.rst
@@ -1,0 +1,9 @@
+Relations Module
+================
+
+.. automodule:: constrainthg.relations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+:doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
