+++ docs/source/research.rst
@@ -1,0 +1,7 @@
+Research
+========
+
+.. include:: chg_overview.rst
+    :start-after: .. start research
+    :end-before: .. end research
+

[…]
