+++ docs/source/demos.rst
@@ -11,9 +11,30 @@ Demonstrations
     3D Printer <https://github.com/jmorris335/Powder-Based-Fusion-Digital-Twin>
     Elevator <https://github.com/jmorris335/ElevatorHypergraph>
 
-.. include:: chg_overview.rst
-    :start-after: .. start demonstrations
-    :end-before: .. end demonstrations
+.. start demonstrations
+
+This is a simple list showing many of the places we've seen CHGs used. If you find more,
+reach out to us to include them here! 
+
+Note that many simple examples are available in the 
+`demos <https://github.com/jmorris335/ConstraintHg/tree/main/demos>`_ directory of the repository.
+
+`Pendulum <https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_pendulum.py>`_
+    Demonstrates model selection.
+
+`Powder Bed Fusion Additive Manufacturing Machine <https://github.com/jmorris335/Powder-Based-Fusion-Digital-Twin>`_
+    Demonstrates interoperability between different part models.
+
+`Elevator <https://github.com/jmorris335/ElevatorHypergraph>`_
+    Combines discrete-event simulation with a PID controller.
+
+`Naval Microgrid <https://github.com/jmorris335/MicrogridHg>`_
+    Complex system featuring data querying and dynamic simulation and linear systems.
+
+`Crankshaft <https://github.com/jmorris335/tool-interoperability-scripts/tree/main>`_
+    Shows how different software can be integrated together such as CAD (Onshape), FEA (Ansys), and MATLAB.
+
+.. end demonstrations
 
 :ref:`genindex` \| :ref:`modindex` \| :ref:`Search <search>`
 
