+++ docs/source/about.rst
@@ -5,14 +5,26 @@
 About
 =====
 
-ConstraintHg is maintained by John Morris, starting with the first commit in October of 2024. The vision is for the package to transform the way we do systems modeling and simulation, enabling interdisciplinary collaboration and understanding of the physical world.
+.. toctree::
+    :hidden:
+
+    contact
+    <Development> contributions
+
+ConstraintHg is maintained by `John Morris <https://orcid.org/0009-0005-6571-1959>`_, starting with the first commit in October of 2024. The vision is for the package to transform the way we do systems modeling and simulation, enabling interdisciplinary collaboration and understanding of the physical world.
 
 Contributions
 -------------
 
-ConstraintHg is intended to be an open-source software guided by its overall purpose. Contributions are welcome subject to review by the developing team. Start by forking the repository.
+.. include:: contributions.rst
+    :start-after: .. start contributions
+    :end-before: .. end contributions
 
-To discuss the package, get help, or otherwise get in contact with the maintenance team, reach out on our `discussion board <https://github.com/jmorris335/ConstraintHg/discussions>`_.
+Get in Touch
+-------------
 
+.. include:: contact.rst
+    :start-after: .. start contact
+    :end-before: .. end contact
 
 :ref:`genindex` \| :ref:`modindex` \| :ref:`Search <search>`
