+++ docs/source/contributions.rst
@@ -1,0 +1,13 @@
+=========================
+Development Contributions
+=========================
+
+.. start contributions
+
+ConstraintHg is intended to be an open-source software guided by its overall purpose. Contributions are welcome subject to review by the developing team. Start by forking the repository, or submitting an issue `here <https://github.com/jmorris335/ConstraintHg/issues>`_.
+
+.. end contributions
+
+To discuss the package, get help, or otherwise get in contact with the maintenance team, reach out on our `discussion board <https://github.com/jmorris335/ConstraintHg/discussions>`_.
+
+:ref:`genindex` \| :ref:`modindex` \| :ref:`Search <search>`
