+++ docs/source/contact.rst
@@ -1,0 +1,15 @@
+=======
+Contact
+=======
+
+.. start contact
+
+Our primary method of communication is our `discussion board <https://github.com/jmorris335/ConstraintHg/discussions>`_ on GitHub. Please use this for discussing software, asking for help, asking about new features, etc. If it's about ConstraintHg, then it should probably go on the discussion board. 
+
+The board also features more theoretical discussions, for instance how the software should be designed, and how we should be working with constraint hypergraphs. Check out our `subtopics <https://github.com/jmorris335/ConstraintHg/discussions/categories/constraint-hypergraph-theory>`_ for more information.
+
+.. end contact
+
+You can also reach out to John directly by emailing jhmrrs AT clemson DOT edu, or by contacting him via his `GitHub profile <https://github.com/jmorris335>`_.
+
+:ref:`genindex` \| :ref:`modindex` \| :ref:`Search <search>`
