### 2025-09-19 0e245f7
Fixed typo in about toctree
