+++ docs/source/about.rst
@@ -9,7 +9,7 @@ About
     :hidden:
 
     contact
-    <Development> contributions
+    Development <contributions>
 
 ConstraintHg is maintained by `John Morris <https://orcid.org/0009-0005-6571-1959>`_, starting with the first commit in October of 2024. The vision is for the package to transform the way we do systems modeling and simulation, enabling interdisciplinary collaboration and understanding of the physical world.
 
