### 2025-10-02 f4d32e0
Don't print cost label for `TNode.__str__` if cost is 0
