### 2025-10-02 #9131 @editorialbot
[REVIEW]: ConstraintHg: A Kernel for Systems Modeling and Simulation
**Submitting author:** @jmorris335 (<a href="http://orcid.org/0009-0005-6571-1959">John Morris</a>)
**Repository:** https://github.com/jmorris335/ConstraintHg
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.1
