### 2025-10-03 33e4cf6
doc: added link to double pendulum demo
