### 2025-10-03 f1ae59d
doc: updated version numbers
