### 2025-10-03 5c8bc8c
doc: typos
