### 2025-10-06 jmorris335/ConstraintHg#27 @fjebaker
Some JOSS review comments and suggestions
Hi @jmorris335! These are some notes to follow up on the review in https://github.com/openjournals/joss-reviews/issues/9131

## General

This is a really cool framework! And a neat way of approaching model simulations using path finding as the means by which to evolve dynamics. But currently I am not entirely convinced of its advertised functionality: it would be useful to have an example that makes use of the automatic model selection that supposedly comes with the BFS over the hypergraph, as described in the "Software Functionalities" section of the paper. The tutorial builds a model of pendulum and then adds a damping sub-model, but it feels like the way you solve such a problem is by either entirely removing one model (the un-damped one) or by writing out every node of the integration over time (expanding the cycle).  These don't 'feel' like viable solutions for an actual modelling system, and doesn't demonstrate the model selection advertised in the overview -- it instead feels like beyond `i = 4`, the bifurcation complexity prevents the method from being practical.

Also the role of cycles in hypergraphs is a little ambiguous at this stage. The paper says that (L 65-68) cycles are allowed and necessary for representing certain real world models, but the documentation says:

> As can be seen from the table, this isn’t just exponential growth, this is
> factorial exponential growth, growing so quickly that after just 8 cycles the
> number of paths outnumbers the number of atoms in the universe by nearly 8 to
> 1. Though pruning strategies exist, the best way to handle this is to bite the
> bullet and to avoid cycles with competing models. If you’re building a
> hypergraph for model selection, then keep your cycles short or expand them out
> explicitly like in Figure 1 <chg_simple>. If you need high-index count cycling,
> then turn off competing edges so that you don’t generate multiple possible
> values for the variables in the cycle.

So cycles are permitted, but cannot be too complex? It would be good to emphasise or state this clearer in the paper.

Then a question for my own curiosity: at the moment, it seems ConstraintHg is single threaded? Could the graph traversal / evaluation be parallelised, e.g. leveraging task graph systems like Dask?

## Paper

- Statement of Need:

  L 24: maybe contextualise or define what you mean by comprehendable? Is this from a mathematical / computational perspective?

  L 25-27: a quick example of why the bond graphs or stock and flow diagrams are limited would be very useful, and would solidify the need ConstraintHg is solving.

  I think this statement might be a little too general also. Narrowing the scope to model composition and/or optimising the simulation of complex systems would, in my opinion, give this more punch. One feature of ConstraintHg to emphasise here is that nodes and edges can be entirely general (even run other models or solvers!), and so the flexibility of the system to capture any model is really powerful -- something other similar tools cannot provide.

- State of the Field:

  L 29-34: You mention two packages here that you say employ macros to discover mathematical relationships between system entities. The macros are in effect just syntactic sugar, however, and from my understanding still compose a functional DAG to represent the system being modelled. That means there is to some degree a direct correspondence between CHG and these other systems, so the contrast is not immediately obvious.

  Is one of the key differences that the hypergraphs are like a meta-graph that can be used to solve meta-problems related to model specifications or composition, and for traversing how these models can be solved? Whereas other generalised simulation frameworks still ultimately have a single pre-defined path through the simulation, yours can discover that path through a BFS?

  Are there other softwares that do similar things? ModelingToolkit.jl comes to mind, though it performs other types of structural simplifications instead of explicit graph traversals, and places more rigorous constraints on what exactly a modelling component can be.

## Documentation

- Tutorial:

  Some broken LaTeX (`\coloneq`), and unresolved links
  - `Figure 2 <chg_nonsimple>`
  - `Figure 1 <chg_simple>`

  The tutorial code does not actually run if copied from the documentation.  There are a few syntax errors and then some other things that differ from the code in the demo.

  - `alpha_val` is undefined.
  - `addEdge` and `printPaths` instead of `add_edge` and `print_paths`.

  Suggestion: The `print_paths` function doesn't actually print to stdout, which is a little misleading. Perhaps make it a `summary()` method and then do `print(hg.summary())`?

  Also the ASCII art is really nice and it should be included in the docs so that a user can see what it should do before they even run it! I notice it is in the README but would be nice to see it in the docs too.

  From looking at the code, it seems you need to specify `*args` and `**kwargs` because all of the relations insert their operations into the first element of `*args`. That might be good to put explicitly in the documentation, both as justification and also since it could be useful for a user to know.

  The hyperlink to `Overview` redirects to the Wiki instead of the `readthedocs` compiled version.

- Guide to CHGs:

  This is a really nice section of the documentation.

  Figure 2 (and subsequent): should that be `-g/r` in the node? Also the equation in the text has `g/m` instead of `g/r`.

Perhaps consider adding a [CONTIRBUTING.md](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributingmd-file) file that formalises some of the instructions in the 'About' page?

## Code

I don't understand the purpose of `main_cli.py`? It seems like ConstraintHg is a library and not a CLI tool?

- Demos:

  `demo_pendulum.py` and `demo_basic.py` run as expected.

  The elevator demo takes a very long time to run (so long I gave up). Some indicator of progress or a simple print that says "takes > 15 minutes" would be useful, as else I do not know if my experience is exceptional or whether I got stuck in a loop.

  For `demo_linear_motion.py` I get an error:
  ```
  Traceback (most recent call last):
    File "/home/fergus/developer/other/ConstraintHg/demos/demo_linear_motion.py", line 22, in <module>
      hg.solve(xn, to_print=True)
      ~~~~~~~~^^^^^^^^^^^^^^^^^^^
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 1393, in solve
      raise e
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 1383, in solve
      t = pf.search(
          min_index=min_index,
      ...<2 lines>...
          search_depth=search_depth,
      )
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 845, in search
      self.explore(root, debug_nodes, debug_edges)
      ~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 873, in explore
      pt = self.make_parent_tnode(combo, edge.target, edge)
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 900, in make_parent_tnode
      parent_t = TNode(label,
                       node_label,
      ...<3 lines>...
                       gen_edge_label=gen_edge_label,
                       gen_edge_cost=edge.weight)
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 141, in __init__
      self.cost = cost
      ^^^^^^^^^
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 202, in cost
      self.calc_cost = self.get_tree_cost()
                       ~~~~~~~~~~~~~~~~~~^^
    File "/home/fergus/developer/other/ConstraintHg/src/constrainthg/hypergraph.py", line 224, in get_tree_cost
      total_cost += self.get_tree_cost(st, checked_edges)

    [Previous line repeated 990 more times]
  RecursionError: maximum recursion depth exceeded
  ```

It would probably be a good idea to include the demos in your test suite or as part of your CI :)
