### 2025-10-10 jmorris335/ConstraintHg#36 @GJHSimmons
[JOSS Review] Tutorial page
[RE: openjournals/joss-reviews/issues/9131]

Hi @jmorris335, I've had time to test the install and have a little play with the package.

I've focused on testing some demos and reading the documentation this week. In the next week or two I hope to handle the paper and other features (e.g. tests).

### Tutorial Page

I started trying to follow the [tutorial page](https://constrainthg.readthedocs.io/en/latest/constrainthg_intro.html).

I found the following errors, where line numbers refer to those in the [source file](https://github.com/jmorris335/ConstraintHg/blob/main/docs/source/constrainthg_intro.rst).

- Line 52: 2x failed `\coloneq` LaTeX commands.
- Line 68: Missing tab before `return` statement.
- Line 86: Maybe make it clearer that if there's just one source node `n_1`, you don't need to enclose it in a list in `add_edge()`.
- Line 126: command is written `addEdge` rather than `add_edge`. I saw this again in Line 179.

In general this page looks outdated when compared to the [pendulum example code](https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_pendulum.py), so I stopped reading here. I'd suggest to update this page, and then I will read through it again.
