### 2025-10-10 jmorris335/ConstraintHg#37 @GJHSimmons
[JOSS Review] Add a Code of Conduct
[RE: openjournals/joss-reviews/issues/9131]

For third-party contributors of open-source software, it is good practice to include a Code of Conduct be included to let people know what behaviour is expected in any interactions.

You can add this directly to your repository, for example by using [this guide](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project).

I'd also recommend to add it as a page within your documentation as well.
