### 2025-10-10 jmorris335/ConstraintHg#38 @GJHSimmons
[JOSS Review] First comments on Demos
[RE: openjournals/joss-reviews/issues/9131]

I've gone through some demos. Please see my comments in the following bullets.

- [Simple demo](https://constrainthg.readthedocs.io/en/latest/simple_demo.html): seems to work great. I believe, however this is the same example as on the [Quickstart](https://constrainthg.readthedocs.io/en/latest/quickstart.html) page. For maintenance reasons, it might be worth linking to this example on the Quickstart page, rather than having two copies of it.
- [Pendulum](https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_pendulum.py): seems to work great.
- [Elevator](https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_elevator.py): eventually runs fine, taking up to the 45 minutes I left it over lunch! It would be good to document better what `helper.py` does and why it is needed.
- [Double pendulum](https://github.com/jmorris335/DrivenPendulum): Works well and quickly.

I have not yet run the other demos. I will let you know if I find any issues when I do.

- In general, the demos included in the main repo [here](https://github.com/jmorris335/ConstraintHg/tree/main/demos) should have more documentation or description, either in themselves or on the documentation website.
- There is also a "Linear Motion" demo I found [here](https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_linear_motion.py). When I try to run this, I obtain:

```
Traceback (most recent call last):
  File "c:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\demo_linear_motion.py", line 22, in <module>
    hg.solve(xn, to_print=True)
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 1393, in solve
    raise e
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 1383, in solve
    t = pf.search(
        ^^^^^^^^^^
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 845, in search
    self.explore(root, debug_nodes, debug_edges)
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 873, in explore
    pt = self.make_parent_tnode(combo, edge.target, edge)
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 900, in make_parent_tnode
    parent_t = TNode(label,
               ^^^^^^^^^^^^
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 141, in __init__
    self.cost = cost
    ^^^^^^^^^
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 202, in cost
    self.calc_cost = self.get_tree_cost()
                     ^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 224, in get_tree_cost
    total_cost += self.get_tree_cost(st, checked_edges)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 224, in get_tree_cost
    total_cost += self.get_tree_cost(st, checked_edges)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 224, in get_tree_cost
    total_cost += self.get_tree_cost(st, checked_edges)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  [Previous line repeated 988 more times]
  File "C:\Users\georg\IDEMS\Software development\IDEMS\JOSS Reviews\ConstraintHG\venv\Lib\site-packages\constrainthg\hypergraph.py", line 222, in get_tree_cost
    checked_edges.add(root.gen_edge_label)
RecursionError: maximum recursion depth exceeded while calling a Python object
```
