+++ docs/source/index.rst
@@ -12,9 +12,9 @@
    :includehidden:
 
    quickstart
-   Tutorial <constrainthg_intro>
+   Tutorial <tutorial/tutorial>
    API <api/index>
-   about
+   about/about
    Demos <demos>
    Repository <https://github.com/jmorris335/ConstraintHg>
 
@@ -22,9 +22,9 @@
    :caption: Constraint Hypergraphs
    :hidden:
 
-   Overview <chg_overview>
-   Guide to CHGs <chg_intro>
-   research
+   CHG Overview <CHGs/chg_overview>
+   Learn About CHGs <CHGs/chg_intro>
+   CHGs/research
 
 .. include:: ../../README.md
    :start-after: ## About
@@ -44,10 +44,10 @@ For a walkthrough, check out the `tutorial <constrainthg_intro>`_ or `demonstrat
 Quick Links
 -----------
 
-Demonstration of package: :doc:`constrainthg_intro`
+Demonstration of package: :doc:`tutorial/tutorial`
    Provides more indepth discussion on constraint hypergraphs and package usage.
 
-Overview of Constraint Hypergraphs: :doc:`chg_overview`
+Overview of Constraint Hypergraphs: :doc:`CHGs/chg_overview`
    Launchpad for resources for learning about Constraint Hypergraphs.
 
 Repository: `<https://github.com/jmorris335/ConstraintHg>`_
