+++ docs/source/tutorial/modeling.rst
@@ -1,0 +1,13 @@
+Modeling
+========
+
+ConstraintHg is a model *organizer*. That means it can take in a given model
+
+Variables
+---------
+
+Edges
+-----
+
+Printing
+--------
