+++ docs/source/tutorial/simulation.rst
@@ -1,0 +1,11 @@
+Simulation
+==========
+
+
+.. figure:: https://github.com/user-attachments/assets/cb8387cc-e005-4ed9-9247-2599f76f323b
+    :alt: Animation of solving a constraint hypergraph
+    :width: 681px
+    :align: center
+    :name: animationfig
+
+    *Visualization of a constraint hypergraph being solved by ConstraintHg*
