+++ docs/source/tutorial/tutorial.rst
@@ -1,0 +1,22 @@
+=====================
+ConstraintHg Tutorial
+=====================
+
+ConstraintHg is a tool for simulating systems. It provides tools for representing and simulating `constraint hypergraphs <../CHGs/chg_overview>`_. It can take in models from across different domains, use cases, and representation schemes, integrating them into a single, queryable structure. Once organized, ConstraintHg allows modelers to aggregate systems (for systems of systems modeling), perform model selection, and, most important of all, autonomously build simulations.
+
+The following tutorial will walk you through how to use ConstraintHg by building a model of a simple pendulum. Note that the full script is available `here <https://github.com/jmorris335/ConstraintHg/blob/main/demos/demo_pendulum.py>`_. For more advanced examples, check out the `Demonstrations <../demos>`_ page. Questions? Get in touch with us on our `discussion board <about/contact>`_!
+
+Tutorial Steps
+--------------
+
+.. toctree::
+   :maxdepth: 1
+   :numbered:
+
+   setup 
+   modeling
+   simulation
+   viability
+   cycles
+   advanced
+
