### 2025-10-11 f97ab26
Finished modeling guide
