+++ docs/source/tutorial/viability.rst
@@ -1,2 +1,5 @@
 Conditional Viability
 =====================
+
+Edge Weights
+------------
