+++ docs/source/tutorial/modeling.rst
@@ -1,13 +1,125 @@
 Modeling
 ========
 
-ConstraintHg is a model *organizer*. That means it can take in a given model
+ConstraintHg is a model *organizer*. That means it can take a given model and structure it so that other models can connect to it without information or meaning being lost. However, you still have to do all the work of providing the model. It is often more convenient to do this work in a different framework. The pendulum example, for instance, can be modeled using Newtonian physics, or setting up the standard Euler-Langrange equations.
 
-Variables
----------
+.. note:: If you want to see a discussion of modeling the pendulum dynamics, check out the overview :ref:`here <pendulum_example>`.
+
+In the linked example, we derive the following constraint hypergraph, which we'll now setup in the Python script.
+
+.. figure:: https://github.com/user-attachments/assets/5512efdb-629c-46e7-9ac4-f919afc1442e
+    :alt: CHG of a pendulum
+    :width: 883px
+    :align: center
+    :name: chg_pend_intro
+    
+    *Constraint hypergraph of the pendulum.*
+
+The first thing to do is to initialize the hypergraph:
+
+.. code-block::
+
+    hg = Hypergraph()
+
+There are a few options you can call when initializing this, such as setting up logging. More information is available at the :ref:`API <hypergraph_init>`. Now we can start passing in the variables.
+
+Nodes
+-----
+
+The constraint hypergraph has several variables which we'll want to include in our model. Think of a the set of variables as all the information we might want to know about the pendulum. Each variable becomes a node in the graph.
+
+.. code-block:: python
+
+    g = hg.add_node("gravity", -10)
+    r = hg.add_node("radius", 0.5)
+    theta0 = hg.add_node("theta0", 3.14159/6)
+    theta = hg.add_node("theta")
+    d_theta = hg.add_node("delta theta")
+    s_theta = hg.add_node("sine theta")
+    F = hg.add_node("gravitational force")
+    omega0 = hg.add_node("omega0", 0.0)
+    omega = hg.add_node("omega")
+    d_omega = hg.add_node("delta omega")
+    c = hg.add_node("damping coeff", 1.5)
+    alpha = hg.add_node("alpha")
+    time_step = hg.add_node("time_step", .03)
+
+The ``hg.add_node()`` method returns a :ref:`Node <node_class>` object, which we can use later to set up the edges in the graph.
+
+In the above we also defined some intermediate notes, these are variables that didn't show up in the original equations, but which will show up in the graph. This is because algebra often condenses multiple operations into a single equation, though for the hypergraph its better to break these operations out so that we can interface with the hidden variables. 
 
 Edges
 -----
+Each edge in a constraint hypergraph maps the values of its source set to the value of its target node. Mathematically this is just function mapping, though you can think of it as assigning a value of a node to every ordered pair of another set of variables. For example, given nodes :math:`A, B := \lbrace 1, 2 \rbrace`, and :math:`C := \lbrace -1, 0, 1 \rbrace`, we could define a hyperedge such that:
+
+======  ======  ======
+ **A**   **B**   **C**
+======  ======  ======
+   1       1       0 
+   1       2       -1 
+   2       1       1 
+   2       2       0
+======  ======  ======
+
+Note that every possible pair :math:`A` and :math:`B` need to be mapped to a corresponding value in :math:`C`. We'll see later on how to take a subset of values  in the relation only. Such a method could encode this table, or we could just note that :math:`C = A - B`. To pass this relationship into the hypergraph, we would just need to pass along a method that took in a value of :math:`A`, a value of :math:`B`, and returned the correct value of :math:`C`. 
+
[…]
