### 2025-10-11 d86c09f
typo in `Hypergraph.__str__`
