### 2025-10-11 2031bef
Closes issue #40
