### 2025-10-11 92fe0c0
Added cycles tutorial
