+++ docs/source/tutorial/viability.rst
@@ -1,5 +1,99 @@
 Conditional Viability
 =====================
 
-Edge Weights
+Intuitive Understanding
+-----------------------
+A function is an algebraic relation that maps every value of one set (the domain) to a value in another set (the codomain). We usually represent our domain set as a combination of variables. However, sometimes, we don't have a relationship for every value in an input set.
+
+For example, let's take our pendulum, which says that :math:`\alpha = -\frac{g}{l}\sin\theta`. That means we have a mapping to :math:`\alpha` for every combination of the variables :math:`g`, :math:`l`, and :math:`\theta`. But that's not *totally* true--if :math:`l` is 0, then our equation is undefined. In other words we don't know what :math:`\alpha` is when the radius of our pendulum is zero!
+
+Via Statements
+--------------
+How do we represent this in our model? What we're trying to do is say that our equation is only applicable for a subset of values in the function's domain. But rather than creating a whole new set of variables to represent :math:`g`, :math:`l`, and :math:`\theta` except for when :math:`l` is 0, we can add what's called a ``via`` statement.
+
+A :ref:`via <edge_init_method>` statement is a boolean method that takes in the values for the edge's source nodes, and returns ``True`` if the values are part of the relations valid subdomain. ``via`` statements are typically given using a ``lambda`` method. 
+
+.. hint:: If you're not familiar with ``lambda`` methods, you can read more about them `here <https://www.w3schools.com/python/python_lambda.asp>`_.
+
+For the example above, we might write a lambda method as the following:
+
+.. code-block:: python
+
+    via=lambda l, *args, **kwargs : l != 0
+
+Note that the arguments for a ``via`` statement are the dict keys of the source nodes passed to an edge.
+
+.. important:: ``via`` statements act the same as relations, meaning that they have to have ``*args, **kwargs`` given as arguments.
+
+Conditional viability is a powerful concept that allows us to create relationships between only a subset of values in a variable. For instance, say you have a variable for whether a door is open, with possible values ``['is_open', 'is_closed']``, and another variable specifying whether the door is locked, with values ``['is_locked', 'is_unlocked']``. You want to indicate that the door is always unlocked if it is open. But you don't necessarily know if the door is locked if it's closed. In order to add the edge, you'll need to add a viability method that resolves to true only if the door is open. Such an edge might look like: ``lambda door_status, *args, **kwargs : door_status == 'is_open'``. Then, if the door is not open, the edge will not be solved.
+
+Psuedo Nodes
 ------------
+
+Sometimes you want to use a property of a node in a ``via`` statement (or perhaps in the main relation). This is tricky since node properties update live during a simulation run. To get around this, you can call what's called a **pseudo node**: a fake node that gets passed to a relation and whose value is a property of some source node for the edge.
+
+To create a pseudo node, we pass a length 2 tuple to the sources of an edge. The first value of the tuple is the key for the source node we're getting the property of, and the second value is the name of the property.
+
+.. important:: For a source node to have a key means that it was passed as a dictionary. This means to use pseudo nodes you *have* to pass the source nodes to the edge in a dict structure.
+
[…]
