+++ docs/source/tutorial/simulation.rst
@@ -1,6 +1,74 @@
 Simulation
 ==========
 
+**Simulation** is the act of artificially establishing information about a system. It's synonyms are prediciton, inference, estimation, and decision-making. Performing a simulation is the act of transforming a set of observed values (inputs) into an unobserved, unknown variable (the output).
+
+Running Our First Simulation
+----------------------------
+
+Usually we create simulations *imperatively*, by taking our inputs and specifying a procedure to transform them into the output. But a constraint hypergraph allows these procedures to be discovered automatically, allowing high-level *declarative* modeling. Each path in a hypergraph represents a possible simulation we can run. To setup a simulation, we use the pathfinding algorithms in ConstraintHg to find a path from a set of inputs to our desired output.
+
+Validation
+__________
+
+First, make sure that all the nodes are connected to each other correctly by recursively printing out the hypergraph: 
+
+.. code-block:: python
+
+    print(hg.printPaths('theta'))
+
+You should get the following output:
+
+.. code-block::
+
+    ──theta, index=1, cost=1
+        ├──theta0, index=1
+        ├◯─theta[CYCLE], index=1
+        │  └──theta0, index=1
+        ├◯─omega, index=1, cost=1
+        │  ├◯─omega[CYCLE], index=1
+        │  ├◯─alpha, index=1, cost=4
+        │  │  └──gravitational force, index=1, cost=3
+        │  │     ├◯─sine theta, index=1, cost=1
+        │  │     │  └──theta[CYCLE], index=1
+        │  │     │     └──theta0, index=1
+        │  │     └●─g/r, index=1, cost=1
+        │  │        ├◯─radius, index=1
+        │  │        └●─gravity, index=1
+        │  ├●─time_step, index=1
+        │  └──omega0, index=1
+        └●─time_step, index=1
+
+Solving
+__________
+
+Once you've checked the model, solve for :math:`\alpha` by calling the :ref:`solve <solve_method>` method on the hypergraph class. It really is that easy! Note that you can pass initial values to the `solve` method, but we don't have to since we set these when we first created the Nodes.
+
+.. code-block:: python
+
+    alpha_tnode = hg.solve(alpha_val)
+    print(alpha_tnode)
+
+This returns:
+
+.. code-block::
+
+    alpha=-10, index=2, cost=5
+
+The object returned by ``Hypergraph.solve`` is a :ref:`TNode <tnode_class>`, which stands for tree node. It recursively defines the simulation path ending at :math:`\alpha`, and allows us to trace back out the simulation. For instance, calling ``alpha_tnode.print_tree()`` returns the simulation:
+
+.. code-block::
+
+    └──alpha=-10, index=2, cost=5
+        └──gravitational force=-10, index=1, cost=4
+            ├──sine theta=0.5, index=1, cost=2
+            │  └──theta=0.5236, index=1, cost=1
+            │     └──theta0=0.5236, index=1
+            └──g/r=-20, index=1, cost=1
+                ├──radius=0.5, index=1
+                └──gravity=-10, index=1
+
+This time with values specified for the nodes. To find this path, the solver runs a Breadth-First Search, which starts from all the inputs defined for the system and iteratively searches all possible paths until one is found leading to the output. This is shown in the animation below:
 
 .. figure:: https://github.com/user-attachments/assets/cb8387cc-e005-4ed9-9247-2599f76f323b
     :alt: Animation of solving a constraint hypergraph
@@ -9,3 +77,32 @@ Simulation
     :name: animationfig
 
     *Visualization of a constraint hypergraph being solved by ConstraintHg*
+
[…]
