+++ docs/source/tutorial/tutorial.rst
@@ -18,7 +18,6 @@ Tutorial Steps
    cycles
    viability
    simulation
-   advanced
 
 .. figure:: https://github.com/user-attachments/assets/c5a77507-bdca-432a-8d3d-b881bec5f26e
     :alt: Pendulum free body diagram
