+++ docs/source/tutorial/simulation.rst
@@ -146,6 +146,7 @@ This gives us competing models: two different ways to solve for :math:`\alpha`. 
 Because we added the disposable tag, our solver can handle the complexity of the diverging branches. Let's try running a simulation over 3 seconds. If you plot the found :math:`\theta` values against time, you should get a plot similar to the following:
 
 .. code-block:: python
+
     theta_tnode = hg.solve(theta, min_index=170)
     thetas, omegas = theta_tnode.values['theta'], theta_tnode.values['omega']
     time = hg.solve(time, min_index=170).values['time']
@@ -169,4 +170,177 @@ Because we added the disposable tag, our solver can handle the complexity of the
 Logging
 -------
 
-Now that we can run simulations, the next step is debugging. 
+Now that we can run simulations, the next step is debugging. It can be tricky to debug a constraint hypergraph. Traditional debugging methods are less effective since we're composing the functions inside a larger library. The best way to see what's going on during pathfinding is to use the logging options for ConstraintHg.
+
+.. note:: ConstraintHg uses the default logging library. To learn how logging works, check out this tutorial `here <https://docs.python.org/3/howto/logging.html#basic-logging-tutorial>`_.
+
+Enabling Logging
+________________
+
+By default, logging is turned off. There are two places that you can turn it on though: when initializing the :ref:`Hypergraph <hypergraph_class>` and when calling the :ref:`Hypergraph.solve <solve_method>` method. The preferred way to setup logging is to set the `logging level <https://docs.python.org/3/library/logging.html#logging-levels>`_. Logging levels are an enumerated type corresponding to a number. The lower the logging level, the more verbose the outputted log.
+
+You can pass either the enumerated type or a plain integer to ConstraintHg to setup logging. To set the logging level to 20 (INFO) you might try:
+
+.. code-block:: python
+
+    from logging import INFO
+    hg = Hypergraph(logging_level=logging.INFO)
+
+.. code-block:: python
+
+    hg.solve(<output>, logging_level=20)
+
+If you setup the logger as an argument to ``Hypergraph.solve`` then the software logs only for that single simulation run, before resetting to whatever the logging level was before the simulation call.
+
+Logging Levels
+______________
+
+The five enumerated logging levels are (from the `docs <https://docs.python.org/3/library/logging.html#logging-levels>`_):
+    - logging.NOTSET: 0
+    - logging.DEBUG: 10
+    - logging:INFO: 20
+    - logging.WARNING: 30
+    - logging.ERROR: 40
+    - logging:CRITICAL: 50
+
+The level you pass determines how verbose the log output will be. ConstraintHg logs on the following levels, with higher levels include all items logged on lower ones:
+    - logging.DEBUG (10): all edges and found combinations are listed, as well as search trees at each explored node.
+    - logging.DEBUG+1 (11): debugging report is logged after a search is complete.
+    - logging.DEBUG+2 (12): edges passed to `debug_edges` and nodes passed to `debug_nodes` as arguments to `Hypergraph.solve` are logged, as well as search trees at each explored node.
+    - logging.INFO (20): start and end of a search are logged.
[…]
