### 2025-10-12 7a1ca0e
Closes issue #36
