+++ demos/demo_pendulum.py
@@ -1,121 +1,174 @@
 from constrainthg.hypergraph import Hypergraph, Node
 import constrainthg.relations as R
-import matplotlib.pyplot as plt
 
-#########################
-### Nodes ##########
-#########################
-
+# Define hypergraph
 hg = Hypergraph()
 
-# -- Nodes
-g = Node("gravity", -10)
-r = Node("radius", 0.5)
-theta0 = Node("theta0", 3.14159/6)
-theta = Node("theta")
-d_theta = Node("delta theta")
-s_theta = Node("sine theta")
-F = Node("gravitational force")
-omega0 = Node("omega0", 0.0)
-omega = Node("omega")
-d_omega = Node("delta omega")
-c = Node("damping coeff", 1.5)
-alpha = Node("alpha")
-time_step = Node("time_step", .03)
+#########################
+### Nodes ###############
+#########################
 
-time = Node("t")
+g = hg.add_node(Node('gravity', -10))
+r = hg.add_node(Node('radius', 0.5))
+theta0 = hg.add_node(Node('theta0', 3.14159/6))
+theta = hg.add_node(Node('theta'))
+d_theta = hg.add_node(Node('delta theta'))
+s_theta = hg.add_node(Node('sine theta'))
+F = hg.add_node(Node('gravitational force'))
+omega0 = hg.add_node(Node('omega0', 0.0))
+omega = hg.add_node(Node('omega'))
+d_omega = hg.add_node(Node('delta omega'))
+c = hg.add_node(Node('damping coeff', 1.5))
+alpha = hg.add_node(Node('alpha'))
+time_step = hg.add_node(Node('time_step', .03))
+time = Node('time', 0.)
 
 ##########################
-##### Relations ##########
+### Relations ############
 ##########################
 
-# For complex systems, it is recommended to define these elsewhere and import
-# See microgrid
+def Rintegrate(step, slope, initial_val, *args, **kwargs):
+    """First order Eulerian integrator."""
+    return step * slope + initial_val
 
-def integrate(s1, s2, s3, **kwargs):
-    return s1 * s3 + s2
+##########################
+### Edges ################
+##########################
 
-############################
-######Edges ################
-############################
+hg.add_edge(
+    sources=theta0,
+    target=theta,
+    rel=R.Rmean,
+    label='theta0->theta',
+)
 
-# Edges from Previous Pendulum Example, but they are updated to current
-#best practice
+hg.add_edge(
+    sources=omega0,
+    target=omega,
+    rel=R.Rmean,
+    label='omega0->omega',
+)
 
+# In case you are wondering why a dictionary is used here instead of a
+# list or set, it is because the ordering of the arguments matters for
+# this input equation g/r, and r/g are very different quantities. 
+hg.add_edge(
+    sources={'s1': g, 's2': r},
+    target='g/r',
+    rel=R.Rdivide,
+    label='(g,r)->b1',
+)
 
-hg.add_edge(theta0, 
-            target= theta, 
-            rel=R.Rmean, 
-            label='theta0->theta')
+hg.add_edge(
+    sources=theta,
+    target=s_theta,
+    rel=R.Rsin,
+    label='theta->sine',
+)
 
-hg.add_edge(omega0, 
-            target=omega, 
-            rel=R.Rmean, 
-            label='omega0->omega')
 # Note that the target (g/r) is not created in the initial node list.
 # Targets and sources that are not located within the list are added. 
+hg.add_edge(
+    sources={'s_theta': s_theta, 'g/r': 'g/r'},
+    target=F,
+    rel=R.Rmultiply,
[…]
