+++ demos/demo_elevator.py
@@ -1,68 +1,73 @@
 from constrainthg.hypergraph import Hypergraph, Node, Edge
 import constrainthg.relations as R
 
-from helper import plotTimeValues  #For visualization, not necessary for simulation
+import matplotlib.pyplot as plt #For visualization, not necessary for simulation
 
 hg = Hypergraph()
 
 # Nodes
 ## Nodes (variables) for elevator stops
-floor = Node('floor', description='the number of a floor', units='floor')
-destination = Node('destination floor', super_nodes=[floor], description='floor number of destination', units='floor')
-curr_floor = Node('current floor', super_nodes=[floor], description='current_floor', units='floor')
-gap = Node('interfloor height', 10., description='height of a single floor', units='m')
-floor_height = Node('height of floor', description='the height of a given floor', units='m')
-dest_height = Node('destination height', super_nodes=[floor_height], description='the height of the destination floor', units='m')
-height_tolerance = Node('height tolerance', 10, description='tolerance on measuring height', units='m')
+floor = hg.add_node(Node('floor', description='the number of a floor', units='floor'))
+destination = hg.add_node(Node('destination floor', super_nodes=[floor], description='floor number of destination', units='floor'))
+curr_floor = hg.add_node(Node('current floor', super_nodes=[floor], description='current_floor', units='floor'))
+gap = hg.add_node(Node('interfloor height', 10., description='height of a single floor', units='m'))
+floor_height = hg.add_node(Node('height of floor', description='the height of a given floor', units='m'))
+dest_height = hg.add_node(Node('destination height', super_nodes=[floor_height], description='the height of the destination floor', units='m'))
+height_tolerance = hg.add_node(Node('height tolerance', 10, description='tolerance on measuring height', units='m'))
 
 ## Nodes for motor controller
-error = Node('error', description='difference beetween destination and current position', units='m')
-KP = Node('K_P', 271, description='proportional gain for PID')
-KI = Node('K_I', 35, description='integral gain for PID')
-KD = Node('K_D', -0.79, description='derivative gain for PID')
-P = Node('P', description='proportional controller input', units='N')
-I = Node('I', 0.0, description='integrative controller input', units='N')
-D = Node('D', description='derivative controller input', units='N')
-alpha = Node('alpha', 0.5, description='filter constant for low-pass filter')
-error_f = Node('filtered error', 0.0, description='error filtered by low-pass filter', units='m')
-error_f_prev = Node('prev filtered error', description='filtered error from the previous iteration', units='m')
-max_pid = Node('max input', 10000, description='maximum controller input', units='N')
-min_pid = Node('min input', -1000, description='minimum controller input', units='N')
-pid_input = Node('PID input', 0., description='input goverend by PID controller', units='N')
-u = Node('controller input', description='input provided by controller', units='N')
+error = hg.add_node(Node('error', description='difference beetween destination and current position', units='m'))
+KP = hg.add_node(Node('K_P', 271, description='proportional gain for PID'))
[…]
