### 2025-10-13 974d901
Added disposal nodes to elevator
