### 2025-10-14 ea0c6aa
Updated elevator demo, added demo testing
