+++ demos/demo_elevator.py
@@ -1,4 +1,4 @@
-from constrainthg.hypergraph import Hypergraph, Node, Edge
+from constrainthg.hypergraph import Hypergraph, Node, Edge, TNode
 import constrainthg.relations as R
 
 import matplotlib.pyplot as plt #For visualization, not necessary for simulation
@@ -8,12 +8,12 @@ hg = Hypergraph()
 # Nodes
 ## Nodes (variables) for elevator stops
 floor = hg.add_node(Node('floor', description='the number of a floor', units='floor'))
-destination = hg.add_node(Node('destination floor', super_nodes=[floor], description='floor number of destination', units='floor'))
-curr_floor = hg.add_node(Node('current floor', super_nodes=[floor], description='current_floor', units='floor'))
-gap = hg.add_node(Node('interfloor height', 10., description='height of a single floor', units='m'))
-floor_height = hg.add_node(Node('height of floor', description='the height of a given floor', units='m'))
-dest_height = hg.add_node(Node('destination height', super_nodes=[floor_height], description='the height of the destination floor', units='m'))
-height_tolerance = hg.add_node(Node('height tolerance', 10, description='tolerance on measuring height', units='m'))
+destination_floor = hg.add_node(Node('destination_floor', super_nodes=[floor], description='floor number of destination', units='floor'))
+current_floor = hg.add_node(Node('current_floor', super_nodes=[floor], description='current_floor', units='floor'))
+gap = hg.add_node(Node('gap', 10., description='height of a single floor', units='m'))
+floor_height = hg.add_node(Node('floor_height', description='the height of a given floor', units='m'))
+dest_height = hg.add_node(Node('destination_height', super_nodes=[floor_height], description='the height of the destination floor', units='m'))
+height_tolerance = hg.add_node(Node('height_tolerance', 10, description='tolerance on measuring height', units='m'))
 
 ## Nodes for motor controller
 error = hg.add_node(Node('error', description='difference beetween destination and current position', units='m'))
@@ -108,25 +108,25 @@ hg.add_edge(
 ## Hybrid connection relationships
 hg.add_edge(
     sources={'s1': height, 's2': gap, 's3': height_tolerance},
-    target=curr_floor, 
+    target=current_floor, 
     rel=R.Rfloor_divide,
     via=lambda s1, s2, s3, **kwargs : abs(s1 % s2) < s3,
     index_offset=1,
     label='(height,gap,height_tol)->current floor',
 )
 hg.add_edge(
-    sources={'gap':gap, 'floor':floor},
+    sources={'gap': gap, 'floor': current_floor},
     target=floor_height,
     rel=R.Rmultiply,
     disposable=['floor'],
-    label='calc current floor height',
+    label='calc_current_floor_height',
 )
 hg.add_edge(
-    sources={'gap':gap, 'floor':destination},
+    sources={'gap':gap, 'floor':destination_floor},
     target=dest_height,
     rel=R.Rmultiply,
     disposable=['floor'],
-    label='calc current floor height',
+    label='calc destination floor height',
 )
 hg.add_edge(
     sources={'s1':dest_height, 's2':height},
@@ -161,19 +161,19 @@ hg.add_edge(
     index_via=lambda s1, s2, **kw : s1 == s2 + 1,
     edge_props=['DISPOSE_ALL'],
 )
-hg.add_edge(
[…]
