### 2025-10-14 b90dd58
Updated linear motion test typo
