### 2025-10-14 6bbc74f
add matplotlib package import to pytest
