### 2025-10-14 7be01eb
Create CODE_OF_CONDUCT.md
Closes issue #37
