+++ src/constrainthg/relations.py
@@ -24,12 +24,13 @@ limitations under the License.
 Notes:
 ------
 
-- Generally imported as import relations as R
+- Generally imported as ``import relations as R``
 - All relationship functions begin with a capital R, so that they are
-normally called as `R.Rfunction`
-- Each relationship should have `*args`, and `**kwargs` as its
-arguments and only arguments. Specific keywords referenced in kwargs
-should be `s1`, `s2`, ... only.
+  normally called as ``R.Rfunction``
+- Each relationship should have ``*args``, and ``**kwargs`` as its
+  arguments and only arguments. Specific keywords referenced in kwargs
+  should be ``s1``, ``s2``, ... only.
+
 """
 
 import numpy as np
