+++ docs/source/about/about.rst
@@ -10,15 +10,14 @@ About
 
     contact
     Development <contributions>
+    Code of Conduct <conduct>
 
 ConstraintHg is maintained by `John Morris <https://orcid.org/0009-0005-6571-1959>`_, starting with the first commit in October of 2024. The vision is for the package to transform the way we do systems modeling and simulation, enabling interdisciplinary collaboration and understanding of the physical world.
 
 Contributions
 -------------
 
-.. include:: contributions.rst
-    :start-after: .. start contributions
-    :end-before: .. end contributions
+ConstraintHg is intended to be an open-source software guided by its overall purpose. Contributions are welcome subject to review by the developing team. Find more information about how to contribute :doc:`here </about/contributions>`.
 
 Get in Touch
 -------------
@@ -26,5 +25,24 @@ Get in Touch
 .. include:: contact.rst
     :start-after: .. start contact
     :end-before: .. end contact
+
+Development Team
+----------------
+
+ConstraintHg is an open source project. All consistent contributers are noted here.
+
+=================       ======================      ========================
+Name                    Organization                Contributions           
+=================       ======================      ========================
+`John Morris`_          Clemson University          Primary developer        
+`Evan Taylor`_          Clemson University          Revised demonstrations  
+`Fergus Baker`_         Newcastle University        JOSS Reviewer           
+`George Simmons`_       IDEMS International         JOSS Reviewer           
+=================       ======================      ========================
+
+.. _John Morris: https://orcid.org/0009-0005-6571-1959
+.. _Evan Taylor: https://orcid.org/0000-0003-1974-8978
+.. _Fergus Baker: https://sr.ht/~fjebaker/
+.. _George Simmons: https://github.com/GJHSimmons
 
 :ref:`genindex` \| :ref:`modindex` \| :ref:`Search <search>`
