### 2025-10-15 efdd599
Fixed project urls typo
