### 2025-10-21 jmorris335/ConstraintHg#47 @GJHSimmons
[JOSS Review] Paper comments
Paper link: https://raw.githubusercontent.com/openjournals/joss-papers/joss.09131/joss.09131/10.21105.joss.09131.pdf

The paper is written to a high quality, is concise, and includes the necessary sections as specified by [JOSS Guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain). These comments are therefore purely editorial. For small insertions, I've used code blocks to highlight them.

- Line 15: I'm not aware of the term of "glass-boxing" in the context of systems, please either define it or provide a reference.
- Lines 19-20: The first sentence here is quite sweeping. Setting a scene closer to system science rather than "science" as a whole, and defining more carefully what you consider a system, would be more appropriate.
- Line 20: **A** model is **a** description of a system, among potentially many. Suggested rewording: "A model is a description of a system, which represents different elements in a system and how they interact."
- Line 25: Please fix formatting of reference "John Baez et al., 2023" to "Baez et al., 2023".
- Line 31: A space is missing between "relationships" and "(Hedges, 2018)".
- Line 31: "In this context`,` tools such as..."
- Lines 42-43: "...behaviour can be made interpreted holistically." - I assume you don't want "made", but please check.
- Lines 52-55: Your description of imperative simulation is hard to parse on first reading. The claim it requires "explicit procedures to be defined for every combination of inputs to outputs" makes it sound like the maximal claimed number of pairs is always achieved. I'd recommend rephrasing this, and perhaps spending 1-2 lines on a small example, or a continuation of the precipitation example, would be beneficial.
- Line 62: The software functionalities section describes features rather than the space to be filled by the package. Ideally, a JOSS paper does not contain such descriptions or examples, but I am happy for it to exist. Perhaps it should be in its own section, rather than a part of the Statement of Need? The final say on this will be with the editor.

### Other comments

Your references look fine and correctly cited, with the only exception being the John Baez reference, above.

You have mentioned `Psymple` and `Decapodes`, is there other related software that users can use, even more generally in constructing (constraint) hypergraphs, unrelated to the simulation piece? If so, these can go in a "Related Software" section.

Also, it is clear the development of this package is related to active research. If the package is connected to, or supporting, ongoing research, then you may want to include a "Related Research" section.
