### 2025-10-21 a5c0297
Removed HTML centering
