+++ README.md
@@ -2,13 +2,11 @@
 <img src="https://github.com/jmorris335/ConstraintHg/blob/14d9ea2db0e73d440dd4de1491ba0ffee0233d87/media/logo.svg?raw=true" width="300">
 </h1><br>
 
-<p align="center">
 [![DOI](https://zenodo.org/badge/869248124.svg)](https://doi.org/10.5281/zenodo.15278018)
 [![Read the Docs](https://img.shields.io/readthedocs/constrainthg?link=https%3A%2F%2Fconstrainthg.readthedocs.io%2Fen%2Flatest%2Findex.html)](https://constrainthg.readthedocs.io/en/latest/)
 [![GitHub branch check runs](https://img.shields.io/github/check-runs/jmorris335/constrainthg/main?label=test%2Flinter&link=https%3A%2F%2Fgithub.com%2Fjmorris335%2FConstraintHg%2Factions%2Fworkflows%2Fpython-app.yml)](https://github.com/jmorris335/ConstraintHg/actions/workflows/python-app.yml)
 ![GitHub Release](https://img.shields.io/github/v/release/jmorris335/ConstraintHg?include_prereleases&display_name=tag)
 ![GitHub last commit](https://img.shields.io/github/last-commit/jmorris335/ConstraintHg)
-</p>
 
 ## About
 ConstraintHg is a systems modeling kernel written in Python that enables general definition and universal simulation of any system. The kernel breaks a system down into the informational values (nodes) and functional relationships (hyperedges), providing robust simulation through pathfinding operations. This repository is under active development (no official release yet), and is therefore subject to change without warning. **It is not a rigorous data storage solution. Do not use this as a database.**
