### 2025-11-07 jmorris335/ConstraintHg#37 @GJHSimmons
Looks great, all good!

PS: getting around to open issues/finalising my review this/next week after a busy period. Thanks for the patience!
