### 2025-11-12 158e3c9
Clean function ordering
