### 2025-11-13 5566fe3
`args`/`kwargs` in rels; Closes #34
This is not a breaking update, but it does represent a major update to
the style of ConstraintHg, in allowing relations (`rel`, `via`,
`index_via`, etc.) to be expressed without the buffer `*args`, `**kwargs`
arguments. Before this, all relations had to be able to handle any number
of arguments. This made it impractical to use `*args` or `**kwargs` within
a method, and caused several bugs when the boilerplate `**kwargs` wasn't
added.

The fix for this issue is twofold:
1. Transform all passed values for an edge (the sources) into a dictionary,
even if that m
[…]
