### 2025-11-13 4597c34
Removed main_cli
