### 2025-11-13 1aa7da3
Clarify printing methods; Closes #32
This is a breaking commit. The functions that return strings have been
updated to not be called as `print_<thing>`. The methods that were updated
are:
- print_paths -> summary
- print_conn -> get_conn
- print_tree -> get_tree

In addition, relevant documentation, docstrings, examples/demos, and
helper functions have been updated to use the new syntax.

This fix was suggested by @fjebaker
