+++ README.md
@@ -72,7 +72,7 @@ hg.add_edge('F', 'C', R.Rnegate)
 
 We can verify that the hypergraph was made correctly by tracing all possible paths for generating C using the `printPaths` function.
 ```[python]
-print(hg.print_paths('C'))
+print(hg.summary('C'))
 ```
 
 This should give us the following output. Hyperedges are indicated with a `◯`, with the last source separated from other edges with a `●`.
