+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "constrainthg"
-version = "0.2.4"
+version = "0.3.0"
 dependencies = [
     "numpy"
 ]
