### 2025-11-13 f4c62ce
Infinite weights; Closes #51
