### 2025-11-13 bd8aad2
Removed demo.gif from README
