### 2025-11-13 cdd67ac
Addresses #47
