+++ paper/paper.bib
@@ -44,7 +44,6 @@
   abstract = {In physics, Feynman diagrams are used to reason about quantum processes. In the 1980s, it became clear that underlying these diagrams is a powerful analogy between quantum physics and topology. Namely, a linear operator behaves very much like a ‘cobordism’: a manifold representing spacetime, going between two manifolds representing space. This led to a burst of work on topological quantum field theory and ‘quantum topology’. But this was just the beginning: similar diagrams can be used to reason about logic, where they represent proofs, and computation, where they represent programs. With the rise of interest in quantum cryptography and quantum computation, it became clear that there is extensive network of analogies between physics, topology, logic and computation. In this expository paper, we make some of these analogies precise using the concept of ‘closed symmetric monoidal category’. We assume no prior knowledge of category theory, proof theory or computer science.},
   isbn = {978-3-642-12820-2 978-3-642-12821-9},
   langid = {english},
-  keywords = {TOREAD,USEFUL}
 }
 
 @book{borutzkyBondGraphModelling2011,
@@ -70,7 +69,6 @@
   isbn = {978-1-4757-3922-0},
   langid = {english},
   pagetotal = {755},
-  keywords = {INTERESTING,USEFUL}
 }
 
 @article{garicanoWhyOrganizationsFail2016,
@@ -99,7 +97,6 @@
   url = {http://qmro.qmul.ac.uk/xmlui/handle/123456789/23259},
   urldate = {2024-09-10},
   langid = {english},
-  annotation = {Unofficial version: http://www.cs.ox.ac.uk/people/julian.hedges/papers/Thesis.pdf}
 }
 
 @article{hoffmannDestinationEarthDigital2023,
@@ -114,6 +111,20 @@
   doi = {10.1016/j.cliser.2023.100394},
   abstract = {The European Destination Earth (DestinE) initiative aims at implementing a digital twin of the Earth system. DestinE combines high-end physical with impact science enabled by novel digital technologies to create a new, interactive digital information system in support of decision-making. Climate change adaptation and disaster risk management for weather-extremes have been selected as priority targets for the first implementation phase. The digital-twin paradigm prompts co-design of the information system with experts and users, for whom DestinE is creating opportunities in a variety of sectors, among others, the management of renewable resources, water, food, energy, and public health. From the start of the initiative, DestinE aims to build on the interest in existing communities and invites them to contribute and design use cases that embrace novel ways of information sharing.},
   keywords = {Co-design,Decision making,Destination Earth,Digital twins,Earth system models,Green deal}
+}
+
+@article{Landry_XGI_2023,
+    author = {Landry, Nicholas W. and Lucas, Maxime and Iacopini, Iacopo and Petri, Giovanni and Schwarze, Alice and Patania, Alice and Torres, Leo},
+    title = {{XGI: A Python package for higher-order interaction networks}},
+    doi = {10.21105/joss.05162},
+    journal = {Journal of Open Source Software},
+    publisher = {The Open Journal},
+    year = {2023},
+    month = may,
+    volume = {8},
+    number = {85},
+    pages = {5162},
+    url = {https://doi.org/10.21105/joss.05162},
 }
 
 @book{leinsterBasicCategoryTheory2014,
@@ -144,7 +155,36 @@
   doi = {10.18576/amis/110106},
[…]
