+++ paper/paper.md
@@ -22,32 +22,31 @@ bibliography: paper.bib
 ---
 
 # Summary
-ConstraintHg is a systems modeling kernel used for parsing constraint hypergraphs. Constraint hypergraphs are a mathematical formalism embodying the constraint-based approach to representing behavior [@willemsBehavioralApproachOpen2007]. Any executable model--whether database schema, plant controller, or ecological forecaster--can be represented as a constraint hypergraph [@morrisUnifiedSystemModeling2025]. Once combined, the unified structure shows how all elements in the system are related to each other. In addition to representing constraint hypergraphs, ConstraintHg provides methods for traversing them, equivalent to simulating the system. While most system simulations are imperatively defined, ConstraintHg enables simulations to be declaratively constructed. The result is the effective glass-boxing of any system, allowing modelers to observe interactions and complex behaviors simply by identifying elements of interest.
+ConstraintHg is a systems modeling kernel used for parsing constraint hypergraphs. Constraint hypergraphs are a mathematical formalism embodying the constraint-based approach to representing behavior [@willemsBehavioralApproachOpen2007]. Any executable model--whether database schema, plant controller, or ecological forecaster--can be represented as a constraint hypergraph [@morrisUnifiedSystemModeling2025]. Once combined, the unified structure shows how all elements in the system are related to each other. In addition to representing constraint hypergraphs, ConstraintHg provides methods for traversing them, equivalent to simulating the system. While most system simulations must be imperatively defined, ConstraintHg enables simulations to be constructed declaratively. As a result, knowledge about a system can be discovered autonomously, transforming a general system representation into an agentic information provider.
 
 # Statement of Need
-Every thing in the world is a system and science is the work of characterizing their behavior [@cellierContinuousSystemModeling1991]. The description of a system is given by a model, which tells how different elements in a system interact. Models are ubiquitous: they are used to describe phenomena including weather patterns [@hoffmannDestinationEarthDigital2023], economic policies [@garicanoWhyOrganizationsFail2016], ecological events [@wangHowVastDigital2025] and immunological responses [@sahalPersonalDigitalTwin2022]. To be comprehendable, models must adhere to an established formalism such as a bond graph [@borutzkyBondGraphModelling2011] or stock and flow diagram [@baezCompositionalModelingStock2023]. The limitations of these often incompatible frameworks makes integrating models difficult, leaving scientists, engineers, and decision makers unable to represent the interactions between disparate elements.
+Every thing in the world is a system: a collection of distinguishable elements that together espouse some unique behavior [@cellierContinuousSystemModeling1991]. To determine that behavior, scientists describe systems using models, which express how different elements in a system interact. Models are ubiquitous: they are used to describe phenomena including weather patterns [@hoffmannDestinationEarthDigital2023], economic policies [@garicanoWhyOrganizationsFail2016], ecological events [@wangHowVastDigital2025] and immunological responses [@sahalPersonalDigitalTwin2022]. To be comprehendable, models must adhere to an established formalism such as a bond graph [@borutzkyBondGraphModelling2011] or stock and flow diagram [@baezCompositionalModelingStock2023]. The limitations of these often incompatible frameworks makes integrating models difficult, leaving scientists, engineers, and decision makers unable to represent the interactions between disparate elements.
 
 ## Approaches to Generalized System Modeling
[…]
