+++ paper/paper.bib
@@ -33,7 +33,7 @@
   title = {Physics, {{Topology}}, {{Logic}} and {{Computation}}: {{A Rosetta Stone}}},
   shorttitle = {Physics, {{Topology}}, {{Logic}} and {{Computation}}},
   booktitle = {New {{Structures}} for {{Physics}}},
-  author = {Baez, J. and Stay, M.},
+  author = {Baez, John and Stay, M.},
   editor = {Coecke, Bob},
   date = {2010},
   volume = {813},
@@ -186,6 +186,23 @@
   pubstate = {prepublished},
   keywords = {Computer Science - Mathematical Software,Computer Science - Software Engineering,Computer Science - Symbolic Computation}
 }
+
+@article{mattssonModelicaInternationalEffort1997,
+  title = {Modelica - {{An International Effort}} to {{Design}} the {{Next Generation Modeling Language}}},
+  author = {Mattsson, Sven Erik and Elmqvist, Hilding},
+  date = {1997-04-01},
+  journaltitle = {IFAC Proceedings Volumes},
+  shortjournal = {IFAC Proceedings Volumes},
+  series = {7th {{IFAC Symposium}} on {{Computer Aided Control Systems Design}} ({{CACSD}} '97), {{Gent}}, {{Belgium}}, 28-30 {{April}}},
+  volume = {30},
+  number = {4},
+  pages = {151--155},
+  issn = {1474-6670},
+  doi = {10.1016/S1474-6670(17)43628-7},
+  abstract = {A new language called Modelica for physical modeling is developed in an international effort. The main purpose is to make it easy to exchange models and model libraries. The design approach builds on non-casual modeling with true equations and the use of object-oriented constructs to facilitate reuse of modeling knowledge. There are already several modeling language based on these ideas available from universities and small companies. There is also significant experience of using them in various applications. The aim of the Modelica effort is to unify the concepts and design a new uniform model language for model representation. The paper describes the effort and gives an overview of Modelica.},
+  keywords = {differential-algebraic equations,hierarchical systems,Modelica,modeling,modeling language,object-orientation,simulation}
+}
+
 
 @online{morrisConstraintHypergraphsUnifying2025,
   title = {Constraint {{Hypergraphs}} as a {{Unifying Framework}} for {{Digital Twins}}},
