+++ paper/paper.bib
@@ -113,7 +113,7 @@
   keywords = {Co-design,Decision making,Destination Earth,Digital twins,Earth system models,Green deal}
 }
 
-@article{Landry_XGI_2023,
+@article{landryXGIPythonPackage2023,
     author = {Landry, Nicholas W. and Lucas, Maxime and Iacopini, Iacopo and Petri, Giovanni and Schwarze, Alice and Patania, Alice and Torres, Leo},
     title = {{XGI: A Python package for higher-order interaction networks}},
     doi = {10.21105/joss.05162},
