+++ paper/paper.md
@@ -25,24 +25,24 @@ bibliography: paper.bib
 ConstraintHg is a systems modeling kernel used for parsing constraint hypergraphs. Constraint hypergraphs are a mathematical formalism embodying the constraint-based approach to representing behavior [@willemsBehavioralApproachOpen2007]. Any executable model--whether database schema, plant controller, or ecological forecaster--can be represented as a constraint hypergraph [@morrisUnifiedSystemModeling2025]. Once combined, the unified structure shows how all elements in the system are related to each other. In addition to representing constraint hypergraphs, ConstraintHg provides methods for traversing them, equivalent to simulating the system. While most system simulations must be imperatively defined, ConstraintHg enables simulations to be constructed declaratively. As a result, knowledge about a system can be discovered autonomously, transforming a general system representation into an agentic information provider.
 
 # Statement of Need
-Every thing in the world is a system: a collection of distinguishable elements that together espouse some unique behavior [@cellierContinuousSystemModeling1991]. To determine that behavior, scientists describe systems using models, which express how different elements in a system interact. Models are ubiquitous: they are used to describe phenomena including weather patterns [@hoffmannDestinationEarthDigital2023], economic policies [@garicanoWhyOrganizationsFail2016], ecological events [@wangHowVastDigital2025] and immunological responses [@sahalPersonalDigitalTwin2022]. For the behavior of a system to be understandable, models must adhere to an established formalism such as a bond graph [@borutzkyBondGraphModelling2011] or stock and flow diagram [@baezCompositionalModelingStock2023]. Information in such frameworks can be difficult to connect, such as comparing energy in a bond graph to population dynamics in a stock and flow diagram. These domain-specificities makes integrating models difficult, leaving scientists, engineers, and decision-makers often unable to discern the complex, multi-physics interactions.
+Every thing in the world is a system: a collection of distinguishable elements that together espouse some unique behavior [@cellierContinuousSystemModeling1991]. To determine that behavior, scientists describe systems using models, which express how different elements in a system interact. Models are ubiquitous: they are used to describe phenomena including weather patterns [@hoffmannDestinationEarthDigital2023], economic policies [@garicanoWhyOrganizationsFail2016], ecological events [@wangHowVastDigital2025] and immunological responses [@sahalPersonalDigitalTwin2022]. For the behavior of a system to be understandable, models must adhere to an established framework such as a bond graph [@borutzkyBondGraphModelling2011] or stock and flow diagram [@baezCompositionalModelingStock2023]. Information in such frameworks can be difficult to connect, such as comparing energy in a bond graph to population dynamics in a stock and flow diagram. These domain-specificities makes integrating models difficult, leaving scientists, engineers, and decision-makers often unable to discern the complex, multi-physics interactions.
 
 ## Approaches to Generalized System Modeling
 The problem with general systems modeling is often addressed under the auspices of category theory[^1] [@leinsterBasicCategoryTheory2014], which provides the mathematical tools for understanding intersystem relationships [@hedgesCompositionalGameTheory2018]. In this context, tools such as [Psymple](https://casasglobal-org.github.io/psymple/latest/) [@simmonsPsymplePythonPackage2025], [Decapodes](https://algebraicjulia.github.io/Decapodes.jl/dev/) [@morrisDecapodesDiagrammaticTool2024], [ModelingToolkit.jl](https://docs.sciml.ai/ModelingToolkit/stable/) [@maModelingToolkitComposableGraph2022], and [Modelica](https://modelica.org) [@mattssonModelicaInternationalEffort1997] have been created. All of these employ macros to discover mathematical relationships between system entities, allowing for the simulation of dynamic systems.
 
[…]
