### 2025-11-14 jmorris335/ConstraintHg#27 @jmorris335
Hi @fjebaker, sorry for the delay. I think I've finally finished responding to all the points here. In addition to updating the paper as suggested, I've also completed the 7 sub-issues listed. Please let me know if I answered all your questions and if you have any followup feedback. Of course, I'll leave closing this issue up to you.
