### 2025-11-14 jmorris335/ConstraintHg#27 @fjebaker
Hello @jmorris335, thanks for pinging me with the update! I am currently travelling visiting colleagues, so it may take me a couple of weeks to review your updates. I'll ping you back when I make progress :)
