### 2025-11-18 fd6022c
Typos in CHG overview
