+++ docs/source/CHGs/chg_overview.rst
@@ -11,6 +11,8 @@ Good Starting Points
 
 Brief Introduction with examples: :doc:`chg_intro`
 
+<iframe width="560" height="315" src="https://www.youtube.com/embed/Ph2yhaThex0?si=gpxri9GmnVeGeD1B" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+
 `Video <https://www.youtube.com/watch?v=Ph2yhaThex0>`_ overview of using CHGs with digital twins: 
 
 .. raw:: html
@@ -27,17 +29,17 @@ More In-Depth
 Demonstrations
 ==============
 
-.. include:: demos.rst
+.. include:: ../demos.rst
     :start-after: .. start demonstrations
     :end-before: .. end demonstrations
 
 
 Questions
 =========
-Reach out at our `discussion board <https://github.com/jmorris335/ConstraintHg/discussions/categories/q-a>`_ to ask questions and see relevant updates.
+Reach out at our `discussion board <https://github.com/jmorris335/ConstraintHg/discussions/categories/q-a>`__ to ask questions and see relevant updates.
 
 
-:doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
+.. :doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
 
 .. _ConstraintHg Repo: https://github.com/jmorris335/ConstraintHg
 .. _home: https://constrainthg.readthedocs.io/en/latest/
