+++ docs/source/CHGs/research.rst
@@ -6,19 +6,19 @@ how they are being used.
 
 .. start research
 
-Unified Systems Modeling and Simulation: `DOI <https://doi.org/10.1115/1.4068375>`_
+Unified Systems Modeling and Simulation: `DOI <https://doi.org/10.1115/1.4068375>`__
     First published paper covering Contraint Hypergraphs, goes into technical 
-    details about how they are defined. A video overview is available `here <https://www.youtube.com/watch?v=nQaxbmd1yds>`_
+    details about how they are defined. A video overview is available `here <https://www.youtube.com/watch?v=nQaxbmd1yds>`__
 
-Integrating Simulation Software with CHGs: `Linked here <https://www.people.clemson.edu/jhmrrs/publications/#:~:text=Declarative%20Integration%20of%20CAD%20Software%20into%20Multi%2DPhysics%20Simulation%20via%20Constraint%20Hypergraphs>`_
+Integrating Simulation Software with CHGs: `Linked here <https://www.people.clemson.edu/jhmrrs/publications/#:~:text=Declarative%20Integration%20of%20CAD%20Software%20into%20Multi%2DPhysics%20Simulation%20via%20Constraint%20Hypergraphs>`__
     Shows how CHGs enable declarative modeling between simulation software, 
     demonstrating with integration of a dynamic model with Onshape's CAD API. 
-    The code for this model is available `here <https://github.com/jmorris335/tool-interoperability-scripts/tree/main>`_
+    The code for this model is available `here <https://github.com/jmorris335/tool-interoperability-scripts/tree/main>`__
 
-Constraint Hypergraphs as a foundation for Digital Twins: `Preprint <https://doi.org/10.48550/arXiv.2507.05494>`_
+Constraint Hypergraphs as a foundation for Digital Twins: `Preprint <https://doi.org/10.48550/arXiv.2507.05494>`__
     Paper showing how CHGs form a robust framework for working with digital twins, 
     enabling observation and simulation for autonomous agents.
 
 .. end research
 
-:doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
+.. :doc:`Home </index>` \| :ref:`genindex` \| :ref:`Search <search>`
