+++ docs/source/conf.py
@@ -25,6 +25,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx.ext.mathjax',
+    'sphinxcontrib.youtube',
 ]
 
 templates_path = ['_templates']
