### 2025-12-02 0f3a34a
Removed YT extension
