### 2025-12-10 3f98a02
demo: updated pendulum plot to v0.3
