+++ paper/paper.md
@@ -35,7 +35,7 @@ The problem with general systems modeling is often addressed under the auspices 
 Constraint hypergraphs (CHGs), conversely, are a functional representation of a system where the system solution is composed from elements in the model rather than applied by external solvers. This allows systems to be represented without requiring a proscriptive interface. A CHG is a mathematical structure that represent the state variables of a system as nodes in a graph and the behaviors of the system as directed hyperedges mapping between these nodes [@morrisUnifiedSystemModeling2025]. Eeach hyperedge representing a function constraining the value of the target node. The generality of this formalism is able to flexibly capture the intracacies of any system. Consequently, independent models defined across domains, scales, or even different software platforms can all be reconciled into a cohesive structure from which behavior can be interpreted holistically.
 
 ## System Simulation
-A model is only as useful as it is simulatable. Simulation is the use of a model to identify unobserved information. For example, a person observing someoine walk into a building with a wet umbrella might infer that it is raining outside. This inference is the simulation of a mental model associating wet umbrellas with precipitation. Simulation can be understood as a function, mapping a set of known inputs (the wetness of the umbrella) to the unobserved output (the current weather). Traditional system simulation is imperative, where the modeler specifies how the explicit processes by which the transformation should be conducted. Imperative simulations are difficult to adapt; for instance, a model might define relationships mapping wetness to precipitation, cloud cover, transit delay, temperature, etc. Describing these imperatively requires a unique sequence to be explicitly defined for each pairing of inputs to outputs. For a system with $n$ variables, the maximum limit of simulation functions that could be defined is $n \left( 2^{n-1}-1 \right)$. This exponential cap makes it untenable to fully describe all imperative simulations for models with even a moderate number of states. CHGs address this by representing a system graphically such that global simulations can be created by combining local interactions. This drastically reduces the complexity of the model by avoiding redundant redefinitions.
+A model is only as useful as it is simulatable. Simulation is the use of a model to identify unobserved information. For example, a person observing someone walk into a building with a wet umbrella might infer that it is raining outside. This inference is the simulation of a mental model associating wet umbrellas with precipitation. Simulation can be understood as a function, mapping a set of known inputs (the wetness of the umbrella) to the unobserved output (the current weather). Traditional system simulation is imperative, where the modeler specifies how the explicit processes by which the transformation should be conducted. Imperative simulations are difficult to adapt; for instance, a model might define relationships mapping wetness to precipitation, cloud cover, transit delay, temperature, etc. Describing these imperatively requires a unique sequence to be explicitly defined for each pairing of inputs to outputs. For a system with $n$ variables, the maximum limit of simulation functions that could be defined is $n \left( 2^{n-1}-1 \right)$. This exponential cap makes it untenable to fully describe all imperative simulations for models with even a moderate number of states. CHGs address this by representing a system graphically such that global simulations can be created by combining local interactions. This drastically reduces the complexity of the model by avoiding redundant redefinitions.
 
 While many packages are provided for representing hypergraphs, such as XGI [@landryXGIPythonPackage2023] or HyperNetX [@praggastisHyperNetXPythonPackage2024], ConstraintHg is the first library for performing declarative simulation using a CHG. It does this by employing a breadth-first search (BFS) algorithm that can autonomously construct an optimal simulation process between arbitrarily paired input and output variables. In addition to finding simple paths, the search algorithm is able to unravel cycles in the hypergraph by performing parallel searches on cycle branches to ensure an optimal chain is found. This allows ConstraintHg to simlulate dynamic systems, where the states of a system go through some type of repeated change.
 
