### 2025-12-10 1ea8f3e
doc: added reference to fig 2; addresses #27
