### 2025-12-10 67e41d7
doc: updated cycle figures, addresses #27
