### 2025-12-10 jmorris335/ConstraintHg#27 @jmorris335
Hi @fjebaker,

Thanks for taking the time to make such a thorough review--it really helps to have such willing reviewers. I'm fixed the issues that you pointed out doing the following:
- I've updated the cycles in 67e41d7 so that the figures are more clear--as they were there was plenty of room for confusion. Hopefully I've made it clearer that the $i \geq n$ is actually the exit condition for the loop, so that $x_n$ can be properly set as the latest value of $x_i$. I also cleaned up the extra markdown in 1ea8f3e.
- I missed updating the commented out code in the pendulum demo. I've since updated it so that it is inline with v0.3 in 3f98a02. You should be able to uncomment and run it (as long as your ConstraintHg version is up to date with what's on PyPI). Thanks for catching this.
- I left the elevator demo unchanged. Your image is correct, but the problem is a modeling one--as given, the PID controller is terribly tuned and causes massive overshoot. I do need to go through and make a better controller, but I'm considering this nice -to-have not need-to-have, since the elevator demo is intended to show model integration versus the quality of the model. But I'll eventually go through and make this better!
- Thanks for catching the paper typo, I fixed this in 3f2dd2a.

Also, thank you for the video reference, I'm looking forward to watching this and learning more!
