### 2025-12-12 jmorris335/ConstraintHg#47 @GJHSimmons
Just one more thing, is the link suggested by the editorial bot https://doi.org/10.1115/detc2025-168376 now a valid DOI for the reference on line 139?
