### 2025-12-12 54e1e81
Fixed typo, updated reference; addresses #47
