+++ paper/paper.bib
@@ -237,12 +237,15 @@
   booktitle = {Proceedings of the {{ASME}} 2025 {{International Design Engineering Technical Conferences}} and {{Computers}} and {{Information}} in {{Engineering Conference}}},
   author = {Morris, John and Mocko, Gregory and Wagner, John and Ramnath, Satchit},
   date = {2025-08-17/2025-08-20},
+  volume = {45th Computers and Information in Engineering Conference (CIE)},
   publisher = {ASME},
   location = {Anaheim, CA},
+  doi = {10.1115/DETC2025-168376},
   abstract = {Declarative modeling frameworks, such as Modelica, are often used to represent systems that require reusable and interoperable models. Simulation in such a framework requires a solver that is capable of transforming the model into an executable process.  However, most declarative solvers are insular, in that they are unable to integrate with software applications needed for simulating complex systems, limiting their usability for simulating multi-domain models. This paper describes a process for creating declarative models that can integrate with an external tool by deconstructing its Application Programmer Interface (API) into a set of functions arranged into a constraint hypergraph. The constraint hypergraph's solver is shown to be capable of automatically parsing these functions to simulate arbitrary pairs of inputs and outputs. The result is a holistic modeling framework that allows for flexible simulation of a complex system, integrates directly with otherwise sequestered platforms, and reveals cross-cutting interactions between system elements. This is demonstrated by integrating the solid modeling capabilities of Onshape with a dynamic model of a crankshaft from a piston engine, showing how a geometric model can be integrated with independently-defined dynamic models. This validates the framework on a limited scale, setting the foundation for work for fully integrating disparate tools into multi-domain, multi-physics modeling and simulation.},
   eventtitle = {{{ASME IDETC-CIE}} 2025},
-  annotation = {Video: https://www.youtube.com/embed/TZ7xuuvDaRo?si=bTiqNOBir9NP5hF0,}
+  isbn = {978-0-7918-8920-6},
 }
+
 
 @software{morrisMicrogridHg2025,
   title = {{{MicrogridHg}}},
