+++ paper/paper.md
@@ -39,7 +39,7 @@ A model is only as useful as it is simulatable. Simulation is the use of a model
 
 While many packages are provided for representing hypergraphs, such as XGI [@landryXGIPythonPackage2023] or HyperNetX [@praggastisHyperNetXPythonPackage2024], ConstraintHg is the first library for performing declarative simulation using a CHG. It does this by employing a breadth-first search (BFS) algorithm that can autonomously construct an optimal simulation process between arbitrarily paired input and output variables. In addition to finding simple paths, the search algorithm is able to unravel cycles in the hypergraph by performing parallel searches on cycle branches to ensure an optimal chain is found. This allows ConstraintHg to simlulate dynamic systems, where the states of a system go through some type of repeated change.
 
-In addition to dealing with dynamic systems, ConstraintHg also provides mechanisms for dealing with partial models, such as when a modeler can provide relations for only a subset of a state's values. For example, while an umbrella being wet might imply that it is raining, but if the umbrella is not wet, no prediction can be made as to current weather. These partial relations become model validity frames which, while essential to systems modeling, can be difficult to capture in a modeling framework [@malakCharacterizingAssessingValidity2004]. ConstraintHg captures a validity frame as a boolean function which is conditional to the parser including the edge in a simulation path. This feature allows modelers to specify the conditions in which a model can be relied on, an important aspect of collaborative modeling where users of a model are often unfamiliar with its limitations.
+In addition to dealing with dynamic systems, ConstraintHg also provides mechanisms for dealing with partial models, such as when a modeler can provide relations for only a subset of a state's values. For example, while an umbrella being wet might imply that it is raining, if the umbrella is not wet, no prediction can be made as to current weather. These partial relations become model validity frames which, while essential to systems modeling, can be difficult to capture in a modeling framework [@malakCharacterizingAssessingValidity2004]. ConstraintHg captures a validity frame as a boolean function which is conditional to the parser including the edge in a simulation path. This feature allows modelers to specify the conditions in which a model can be relied on, an important aspect of collaborative modeling where users of a model are often unfamiliar with its limitations.
 
 # Related Research
 ConstraintHg is actively being employed in work deriving the mathematical foundations of digital twins [@morrisConstraintHypergraphsUnifying2025], as well as establishing executable digital threads for model-based enterprises [@morrisDeclarativeIntegrationCAD2025]. It has been employed to modeling a [naval microgird](https://github.com/jmorris335/MicrogridHg) [@morrisMicrogridHg2025], an [elevator lift](https://github.com/jmorris335/ElevatorHypergraph) [@morrisUnifiedSystemModeling2025], a kinematically-constrained [crankshaft](https://github.com/jmorris335/tool-interoperability-scripts) [@morrisDeclarativeIntegrationCAD2025], and an [additive manufacturing machine](https://github.com/jmorris335/Powder-Based-Fusion-Digital-Twin).
