### 2025-12-12 jmorris335/ConstraintHg#47 @jmorris335
Hi @GJHSimmons,

Thanks for doing this thorough review and for all of your helpful comments. The bot is right, the above DOI was recently published. I've attached it to the paper and fixed the typo you mentioned above in 54e1e81.
