### 2025-12-12 #9131 @jmorris335
> [@olexandr-konovalov](https://github.com/olexandr-konovalov) I have completed my review checklist. There is one typo I've just raised with the paper in [this issue](https://github.com/jmorris335/ConstraintHg/issues/47) for [@jmorris335](https://github.com/jmorris335) to address, but all my other recommendations have been thoroughly addressed.

I've updated the paper and fixed the two issues caught by @GJHSimmons. Thank you @fjebaker and @GJHSimmons for the detailed and helpful reviews!
