### 2026-01-06 #9131 @jmorris335
Hi @GJHSimmons, just repinging the above request from @olexandr-konovalov to confirm that you're satisfied with all improvements--otherwise let me know and I'm happy to make further revisions.

Thanks, and hope you had a happy new year!
