+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "constrainthg"
-version = "0.3.0"
+version = "0.3.1"
 dependencies = [
     "numpy"
 ]
