### 2026-01-19 #9131 @olexandr-konovalov
@jmorris335 could you please review https://github.com/jmorris335/ConstraintHg/pull/54 ?

In addition to that, https://zenodo.org/records/18190055 has the title ConstraintHg - please update it so that it will be exactly the title of the paper ("ConstraintHg: A Kernel for Systems Modeling and Simulation").

Thanks!
