### 2026-01-19 #9131 @jmorris335
Hi @olexandr-konovalov,

I'm updated the [Zenodo archive](https://zenodo.org/records/18190055) to have the correct title (apologies, I thought this meant the same title as the software, not the paper itself).

I have not yet merged jmorris335/ConstraintHg#54 as I don't usually prefer spaces around em dashes and wanted to double check that this is the requested style for JOSS before adding them. If it is though, I'm happy to merge!
