### 2026-01-19 #9131 @olexandr-konovalov
@danielskatz could you please comment about the use of dashes? In my writing style, I always put spaces around them, otherwise they are rendered like hyphens. Of course, one can perhaps take the phrase enclosed in `--` and enclose it into `(...)` instead.
