### 2026-01-19 b9249ab
Replace em dashes with commas
It seems the double hyphen (--) is replaced with an en dash (U+2013) in pandoc's compiler. Rather than tackle the issue of using spaces or not to avoid the en dash I'm replacing the punctuation with commas.
