### 2026-01-19 #9131 @jmorris335
Hi @olexandr-konovalov and @danielskatz, I'm happy to use parentheses rather than dashes. I added an addition [commit](https://github.com/jmorris335/ConstraintHg/pull/54/commits/b9249ab978496bad7a4942f982701df423ac310a) to the  pull request (jmorris335/ConstraintHg#54) that replaced the em dashes with commas (which seem to fit the voice of the sentence better). Let me know if this resolves it, and apologies for making a mountain out of a molehill!
