### 2026-01-19 #9131 @danielskatz
I agree that emdashes generally shouldn't have spaces around them
