### 2026-01-19 #9131 @olexandr-konovalov
@jmorris335 @danielskatz I see! I have double-checked Ukrainian orthography which requires spaces around dashes - I guess that's where my style of writing comes from. There is also a bit here about smaller spaces: https://en.wikipedia.org/wiki/Whitespace_character#Hair_spaces_around_dashes in English texts. But I like the resulting solution with commas, so we don't need to go into deeper discusssion.
