### 2026-01-19 #9131 @olexandr-konovalov
The 1st one comes from this line: the length of that string is 62, the limit is 32

https://github.com/jmorris335/ConstraintHg/blob/dc0669ea012b4f13e1c55aa277ec75fd70637737/paper/paper.bib#L240

Can't see where the 2nd comes from, but maybe it's a cascading effect.
