### 2026-01-19 #9131 @danielskatz
Typically, at least in US English as I know it, spaces are common are endashes but not emdashes.
