### 2026-01-19 8f8cb97
Removed excessive volume in bib file
