+++ paper/paper.bib
@@ -237,7 +237,7 @@
   booktitle = {Proceedings of the {{ASME}} 2025 {{International Design Engineering Technical Conferences}} and {{Computers}} and {{Information}} in {{Engineering Conference}}},
   author = {Morris, John and Mocko, Gregory and Wagner, John and Ramnath, Satchit},
   date = {2025-08-17/2025-08-20},
-  volume = {45th Computers and Information in Engineering Conference (CIE)},
+  volume = {2A: 45th CIE Conference},
   publisher = {ASME},
   location = {Anaheim, CA},
   doi = {10.1115/DETC2025-168376},
