### 2026-01-19 #9131 @danielskatz
See line 244 of the .bib file.

volume is typically a number. What's in the volume tag here is not really a volume number but something else. Maybe it's part of the booktitle, or maybe it should just be dropped.

For what it's worth, here's what the doi2bib.org service gives for this DOI

```
  series = {IDETC-CIE2025},
  title = {Declarative Integration of CAD Software Into Multi-Physics Simulation via Constraint Hypergraphs},
  url = {http://doi.org/10.1115/DETC2025-168376},
  DOI = {10.1115/detc2025-168376},
  booktitle = {Volume 2A: 45th Computers and Information in Engineering Conference (CIE)},
  publisher = {American Society of Mechanical Engineers},
  author = {Morris,  John and Mocko,  Gregory and Ramnath,  Satchit and Wagner,  John},
  year = {2025},
  month = aug,
  collection = {IDETC-CIE2025}
```

Also, as I look at the bib file, I see lots of extra {{}}s that should be removed. The only extra {}s should be when a proper name or something similar needs to be preserved with the correct case. Otherwise, JOSS's processing should handle the cases in the JOSS style.
