### 2026-01-19 #9131 @jmorris335
Agreed that the volume name is too long, though that's how ASME is citing it in their digital collection [here](https://asmedigitalcollection.asme.org/IDETC-CIE/proceedings/IDETC-CIE2025/89206/V02AT02A007/1225728). Their provided bibtex (not biblatex) is:

```
@proceedings{10.1115/DETC2025-168376,
    author = {Morris, John and Mocko, Gregory and Ramnath, Satchit and Wagner, John},
    title = {Declarative Integration of CAD Software Into Multi-Physics Simulation via Constraint Hypergraphs},
    volume = {Volume 2A: 45th Computers and Information in Engineering Conference (CIE)},
    series = {International Design Engineering Technical Conferences and Computers and Information in Engineering Conference},
    pages = {V02AT02A007},
    year = {2025},
    month = {08},
    abstract = {...},
    doi = {10.1115/DETC2025-168376},
    url = {https://doi.org/10.1115/DETC2025-168376},
    eprint = {https://asmedigitalcollection.asme.org/IDETC-CIE/proceedings-pdf/IDETC-CIE2025/89206/V02AT02A007/7557249/v02at02a007-detc2025-168376.pdf},
}
```

My current compromise is to shorten the volume to ` volume = {2A: 45th CIE Conference},`, though I'm happy to do something different as needed (such as ` volume = {2A},` or similar).
