### 2026-01-19 #9131 @danielskatz
The doi2bib service pulls the metadata that the published registered along with the DOI, so I would use that. It doesn't really matter in some ways though.

Anyhow, please do also remove the extra {{}}s from the bib file, then we can proceed.
