### 2026-01-19 #9131 @danielskatz
Also, months should be three letter abbreviations, not numbers, so maybe just go with the bibtex I provided, as the ASME bibtext seems to have a number of issues.
