+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @report{amesCategoricalTheoryHybrid2006,
   type = {PhD Dissertation},
-  title = {A {{Categorical Theory}} of {{Hybrid Systems}}},
+  title = {A Categorical Theory of Hybrid Systems},
   author = {Ames, Aaron David},
   date = {2006-12-11},
   number = {UCB/EECS-2006-165},
@@ -11,11 +11,11 @@
 }
 
 @inproceedings{baezCompositionalModelingStock2023,
-  title = {Compositional {{Modeling}} with {{Stock}} and {{Flow Diagrams}}},
-  booktitle = {Proceedings {{Fifth International Conference}} on {{Applied Category Theory}}},
+  title = {Compositional Modeling with Stock and Flow Diagrams},
+  booktitle = {Proceedings Fifth International Conference on Applied Category Theory},
   author = {Baez, John and Li, Xiaoyan and Libkind, Sophie and Osgood, Nathaniel D. and Patterson, Evan},
   date = {2023-08-07},
-  series = {Electronic {{Proceedings}} in {{Theoretical Computer Science}}},
+  series = {Electronic Proceedings in Theoretical Computer Science},
   volume = {380},
   eprint = {2205.08373,},
   eprinttype = {arXiv},
@@ -24,15 +24,15 @@
   location = {Glasgow, United Kingdom},
   doi = {10.4204/EPTCS.380.5},
   abstract = {Stock and flow diagrams are widely used in epidemiology to model the dynamics of populations. Although tools already exist for building these diagrams and simulating the systems they describe, we have created a new package called StockFlow, part of the AlgebraicJulia ecosystem, which uses ideas from category theory to overcome notable limitations of existing software. Compositionality is provided by the theory of decorated cospans: stock and flow diagrams can be composed to form larger ones in an intuitive way formalized by the operad of undirected wiring diagrams. Our approach also cleanly separates the syntax of stock and flow diagrams from the semantics they can be assigned. We consider semantics in ordinary differential equations, although others are possible. As an example, we explain code in StockFlow that implements a simplified version of a COVID-19 model used in Canada.},
-  eventtitle = {Fifth {{International Conference}} on {{Applied Category Theory}} ({{ACT2023}})},
+  eventtitle = {Fifth International Conference on Applied Category Theory (ACT2023)},
   keywords = {Computer Science - Logic in Computer Science,Computer Science - Mathematical Software,Mathematics - Category Theory,Quantitative Biology - Populations and Evolution},
   annotation = {EPTCS: 380.5}
 }
 
 @incollection{baezPhysicsTopologyLogic2010,
-  title = {Physics, {{Topology}}, {{Logic}} and {{Computation}}: {{A Rosetta Stone}}},
-  shorttitle = {Physics, {{Topology}}, {{Logic}} and {{Computation}}},
-  booktitle = {New {{Structures}} for {{Physics}}},
+  title = {Physics, Topology, Logic and Computation: A Rosetta Stone},
+  shorttitle = {Physics, Topology, Logic and Computation},
+  booktitle = {New Structures for Physics},
   author = {Baez, John and Stay, M.},
   editor = {Coecke, Bob},
   date = {2010},
@@ -47,8 +47,8 @@
 }
 
 @book{borutzkyBondGraphModelling2011,
-  title = {Bond {{Graph Modelling}} of {{Engineering Systems}}: {{Theory}}, {{Applications}} and {{Software Support}}},
-  shorttitle = {Bond {{Graph Modelling}} of {{Engineering Systems}}},
[…]
