### 2026-01-19 #9131 @jmorris335
In that case, I'll follow the doi2bib since that seems nicer than the publisher's onsite citation info. I've also removed the extraneous double braces, which were a result of a capitalization setting on my Zotero exporter. This is updated in this [commit](https://github.com/jmorris335/ConstraintHg/commit/a2cf8388cb32961bb35f7b10dd7b079b44dbdb6d). The new biblatex is:

```
@inproceedings{morrisDeclarativeIntegrationCAD2025,
  title = {Declarative Integration of CAD Software into Multi-Physics Simulation via Constraint Hypergraphs},
  booktitle = {Volume 2A: 45th Computers and Information in Engineering Conference (CIE)},
  author = {Morris, John and Mocko, Gregory and Wagner, John and Ramnath, Satchit},
  date = {2025-08-17/2025-08-20},
  publisher = {American Society of Mechanical Engineers},
  location = {Anaheim, CA},
  doi = {10.1115/DETC2025-168376},
  abstract = {...},
  eventtitle = {{{ASME IDETC-CIE}} 2025},
  isbn = {978-0-7918-8920-6},
}
```
