+++ paper/paper.bib
@@ -236,7 +236,7 @@
   title = {Declarative Integration of CAD Software into Multi-Physics Simulation via Constraint Hypergraphs},
   booktitle = {Volume 2A: 45th Computers and Information in Engineering Conference (CIE)},
   author = {Morris, John and Mocko, Gregory and Wagner, John and Ramnath, Satchit},
-  date = {2025-08-17/2025-08-20},
+  date = {2025-08-17},
   publisher = {American Society of Mechanical Engineers},
   location = {Anaheim, CA},
   doi = {10.1115/DETC2025-168376},
