### 2026-01-19 #9131 @jmorris335
Got it: the date in the same bib item was spread across two dates. I've edited it now in this [commit](https://github.com/jmorris335/ConstraintHg/commit/ba2cefb458212a461a3932edc77b36fc4730f4cd), hopefully it compiles now!

Original:
```
date = {2025-08-17/2025-08-20},
```

Resulting [XML](https://github.com/openjournals/joss-papers/blob/joss.09131/joss.09131/10.21105.joss.09131.crossref.xml) (lines 235-6):
```xml
<cYear>2025</cYear>
<cYear>2025</cYear>
```

Updated:
```
 date = {2025-08-17},
```
