### 2026-01-19 #9131 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
