### 2026-01-19 #9131 @olexandr-konovalov
> Typically, at least in US English as I know it, spaces are common are endashes but not emdashes.

To be pedantic, I have checked https://uk.wikipedia.org/wiki/%D0%A2%D0%B8%D1%80%D0%B5 and it says that in Ukrainian endash is used without spaces, but emdash is normally surrounded by space.
