### 2026-01-19 #9131 @jmorris335
> [@jmorris335](https://github.com/jmorris335) - please merge [jmorris335/ConstraintHg#55](https://github.com/jmorris335/ConstraintHg/pull/55) for a few final bib changes.

Done, thank you for catching these proper nouns!
