### 2026-01-19 #9131 @danielskatz
Congratulations to @jmorris335 (John Morris) on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you
